cosine contrastive loss

+++ cosine_contrastive_loss.py
@@ -19,5 +19,5 @@ class CosineContrastiveLoss(nn.Module):
     def forward(self, output1, output2, label):
         cos_sim = F.cosine_similarity(output1, output2)
         loss_cos_con = torch.mean((1-label) * torch.div(torch.pow((1.0-cos_sim), 2), 4) +
-                                    (label) * torch.pow(cos_sim * torch.lt(cos_sim, self.margin), 2))
+                                    (label) * torch.pow(cos_sim * torch.lt(cos_sim, self.margin).float(), 2))
         return loss_cos_con
